requirements.txt
spotipy>=2.23.0
python-dotenv>=1.0.0